Fixes plink generating invalid results.
Even though the root cause of the issue is not clear, the program did
not perform the computations correctly with msvc with our rudimentary
port. This turns out to be fixable with using either mingw or LLVM. Both
are tested to work. Consequently, Qt should be configured to use one of
these two backends. (This is a local setting for Qt that is not part of
CMakeLists.txt):