First integrated and fully working version
In this version, we added the plink_monitor panel that handles outputs
from plink. We moved CMakeLists.txt for the main GUI program to the root
folder, as well as modifications to it that enables the plink executable
built in the correct directory with the "build all" option in Qt.